services/bot/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Deprecated==1.2.13
discord.py==2.1.0
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.3
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.13
werkzeug==2.1.2
wrapt==1.14.1
yarl==1.8.2
